+++ migrations/000004_create_email_table.down.sql
@@ -1,0 +1,2 @@
+-- Drop email table
+DROP TABLE IF EXISTS email;